feat: 대기 순번 필드 rank 추가
- 대기 순번을 조회 시점 계산값이 아니라 슬롯별 저장값으로 고정하는 방식으로 수정

+++ src/main/java/roomescape/reservation/application/dto/WaitingDetail.java
@@ -12,5 +12,5 @@ public record WaitingDetail(Long waitingId,
                             String thumbnailImgUrl,
                             Long timeId,
                             LocalTime startAt,
-                            Long rank) {
+                            int rank) {
 }
